+++ lang/en/theme_boost_union.php
@@ -687,6 +687,14 @@ $string['showsitehomerighthandblockdraweronfirstloginsetting_desc'] = 'With this
 $string['showsitehomerighthandblockdraweronguestloginsetting'] = 'Show right-hand block drawer of site home on guest login';
 $string['showsitehomerighthandblockdraweronguestloginsetting_desc'] = 'With this setting, the right-hand block drawer of site home will be displayed in its expanded state by default. This only applies to users who log in as a guest.';
 
+// Settings: Home site tab.
+$string['homesitetab'] = 'Home site';
+// ... Section: Log in link block.
+$string['loginlinkheading'] = 'Log in link';
+// Setting: Display "Login" link as button.
+$string['loginlinkbuttonenabled'] = 'Display "Log in" link as button';
+$string['loginlinkbuttonenabled_desc'] = 'With this setting, you can have the "Log in" link in the top of the page shown as button.';
+
 // Settings: Page layouts tab.
 $string['pagelayoutstab'] = 'Page layouts';
 // ... Section: tool_policy heading.